requirements.txt
loguru>=0.7.2,<0.8.0
pyyaml>=6.0.1,<6.1.0
dataclass-wizard>=0.22.2,<0.23.0
httpx>=0.25.1,<0.26.0
cryptography>=41.0.5,<41.1.0
pydantic>=2.5.2,<2.6.0
jwcrypto>=1.5.0,<1.6.0
flask>=3.0.0,<3.1.0
fastapi>=0.104.1,<0.105.0
uvicorn>=0.24.0,<0.25.0
requests>=2.31.0,<2.32.0
aiohttp>=3.9.0,<3.10.0
simple_term_menu>=1.6.3,<1.7.0
nbclient>=0.9.0,<0.10.0
ipython>=8.18.1,<8.19.0
ipykernel>=6.27.1,<6.28.0
build>=1.0.3,<1.1.0
nbconvert>=7.11.0,<7.12.0